requirements.txt
opencv-python
numpy
Pillow
torch
torchvision
torchaudio
transformers
sentencepiece
tensorflow
tqdm
pyttsx3
vosk
sounddevice
pyaudio
pyzbar
regex
